Micro-service to power off a machine
Sending a POST to /poweroff on the chosen port will run systemctl to
power off the machine. It's still up to the user to figure out how to
make systemd allow the shutdown - so it could be run as root (yikes!),
or possibly something can be worked out with polkit.